New correct commit

+++ app/Http/Controllers/AdminController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Requests\SmileRequest;
 use App\Models\ChatUserModel;
 use App\Models\LikeUser;
 use App\Models\SmileModel;
@@ -25,32 +26,98 @@ class AdminController extends Controller
         return view('admin.main', compact('userMaxLike', 'countMaxLike'));
     }
 
-    public function users()
+    public function users(Request $request)
     {
-        $users = User::get();
+        $getCountry = "";
+        $getTown = "";
+        $getSex = '';
+        $selPar = '';
+        $getAgeMinPar = "";
+        $getAgeMaxPar = "";
+        $getNamePar = '';
+        $getSurnamePar = '';
+        $users = User::query();
+        if($request->get('filter')){
+            $users->orderby('age',$request->get('filter'));
+            $selPar = $request->get('filter');
+        }
 
-        return view('admin.users.index', compact('users'));
+        if($request->get('sex')){
+            $users->where('sex',$request->get('sex'));
+            $getSex = $request->get('sex');
+        }
+
+        if($request->get('town')){
+            $users->where('town',$request->get('town'));
+            $getTown = $request->get('town');
+        }
+
+        if($request->get('country')){
+            $users->where('country',$request->get('country'));
+            $getCountry = $request->get('country');
+        }
+        if($name = $request->get('nameFilled')){
+            $users->where('name','LIKE',"%" . $name . "%");
+            $getNamePar = $request->get('nameFilled');
+        }
+
+        if($request->get('min_age')){
+            $users->where('age','>=', $request->get('min_age'));
+            $getAgeMinPar = $request->get('min_age');
+        }
+
+        if($request->get('max_age')) {
+            $users->where('age', '<=', $request->get('max_age'));
+            $getAgeMaxPar = $request->get('max_age');
+        }
+
+        if($request->get('surnameFilled')) {
+            $users->where('surname', 'LIKE', "%" . $request->get('surnameFilled') . "%");
+            $getSurnamePar = $request->get('surnameFilled');
+        }
+        session()->flash('message', "Найдено {$users->count()} пользователей");
+        $users = $users->paginate(10)
+            ->withPath('?' . $request->getQueryString());
+        $errors = session()->get('changeStatusError');
+        $message = session()->get('message');
+
+
+
+        return view('admin.users.index', compact('users','errors','selPar','message','getNamePar','getSurnamePar','getAgeMinPar','getAgeMaxPar','getSex','getTown','getCountry'));
     }
 
-    public function chats()
+    public function chats(Request $request)
     {
-        $chats = ChatUserModel::get();
-
-        return view('admin.chats.index', compact('chats'));
+        $chats = ChatUserModel::query();
+        $getPar = '';
+        if($request->get('chat_id')){
+            $chats->where('chat_id','=',$request->get('chat_id'));
+            $getPar = $request->get('chat_id');
+        }
+        $chats = $chats->paginate(12)
+            ->withPath('?' . $request->getQueryString());
+        return view('admin.chats.index', compact('chats','getPar'));
     }
 
     public function smiles()
     {
         $smiles = SmileModel::get();
+        $error = session()->get('deleteSmileError');
 
-        return view('admin.smiles.index', compact('smiles'));
+        return view('admin.smiles.index', compact('smiles','error'));
     }
 
-    public function messages()
+    public function messages(Request $request)
     {
-        $messages = \App\Models\Message::get();
-
-        return view('admin.messages.index', compact('messages'));
+        $getPar = '';
+        $messages = \App\Models\Message::query();
+        if($request->get('chat_id')){
+            $messages->where('chat_id','=',$request->get('chat_id'));
+            $getPar = $request->get('chat_id');
+        }
+        $messages = $messages->paginate(12)
+            ->withPath('?' . $request->getQueryString());
+        return view('admin.messages.index', compact('messages','getPar'));
     }
 
     public function likes()
@@ -67,9 +134,9 @@ class AdminController extends Controller
         return $smile;
     }
 
-    public function addSmile(Request $request)
+    public function addSmile(SmileRequest $request)
     {
-        $image = $request->file('image')->store('smile','public');
+        $image = $request->file('content')->store('smile','public');
         $smile = SmileModel::create([
             'content' => $image,
         ]);
@@ -80,6 +147,11 @@ class AdminController extends Controller
     public function deleteSmile(Request $request)
     {
         $smile = SmileModel::find($request->get('id'));
+        if(!$smile)
+        {
+            session()->flash('deleteSmileError','Такого смайлика нету');
+            return redirect()->back();
+        }
         $smile->delete();
 
         return redirect()->back();
@@ -91,5 +163,33 @@ class AdminController extends Controller
         $smile->delete();
 
         return $smile;
+    }
+
+    public function changeStatus(Request $request)
+    {
+        $user = User::find($request->get('user_id'));
+        if(!$user){
+            session()->flash('changeStatusError','Такого пользователя не существует');
+            return redirect()->back();
+        }
+        $user->status = $request->get('user_status');
+        $user->save();
+
+        return redirect()->back();
+    }
+
+    public function deleteAdminChat(Request $request)
+    {
+        $user_sec_id = User::find($request->get('second_id'));
+        $user_first_id = User::find($request->get('first_id'));
+        $chat_id = $user_first_id->chat()->find($user_sec_id)->pivot->chat_id;
+        $messages = \App\Models\Message::get()->where('chat_id','=', $chat_id);
+        if($messages){
+            foreach($messages as $message)
+                $message->delete();
+        }
+        $user_first_id->chat()->detach($user_sec_id->id);
+        $user_sec_id->chat()->detach($user_first_id->id);
+        return redirect()->back();
     }
 }
